Driving Simulation

AutomotiveInteraction.keep_vehicle_subset! — Function

function keep_vehicle_subset!(scene::Scene, ids::Vector{Int})

Just keep a set of vehicles in scene as specified by ids (a list of vehicle ids)

Obtained from ngsim_env/julia/src/ngsim_utils.jl

Example

scene = get_scene(1, traj_interaction)
veh_id_list = [7,10,11,18,22,25,29]
keep_vehicle_subset!(scene,veh_id_list)

Missing docstring.

Missing docstring for place_vehs_on_straight_road. Check Documenter's build log for details.

Missing docstring.

Missing docstring for place_vehs_on_separated_segments_road. Check Documenter's build log for details.

AutomotiveInteraction.make_def_models — Function

function make_def_models(scene)

Takes an input scene and associates a default param Tim2DDriver to all the vehicles in the scene

Examples

scene = get_scene(1,traj_interaction)
models = make_def_models(scene)

AutomotiveInteraction.make_TimLaneChanger_models — Function

function make_TimLaneChanger_models(scene)

Takes input scene and associates default param Tim2DDriver model to all vehs, with lane changer being TimLaneChanger

Examples

scene = get_scene(traj_interaction)
models = make_TimLaneChanger_models(scene)

AutomotiveInteraction.get_hallucination_scenes — Function

function get_hallucination_scenes

Hallucinate starting from start_step for nsteps using models and return a list of scenes

Used by plot_carwise_pos_vel to assess position and velocity traces against ground truth

Returns

halluc_scenes_list: List containing the scenes starting with the ground truth scene at start_step

Examples

scene_list = get_hallucination_scenes(scene,models=models,duration=6)
scenelist2video(scene_list,filename="media/driving_vid.mp4")

AutomotiveInteraction.run_vehicles — Function

function run_vehicles

Select vehicles with ids in id_list from frame number start_frame of track information in traj

Assign default 2D driver model to the vehicle and run it for duration seconds on roadway

Writes a video to location specified by filename

Examples

run_vehicles(id_list=[29,19,28,6,8,25,2,10,7,18,12],roadway=roadway_interaction,filename="media/run_test.mp4")