Roadway Building

These functions are required to build the roadway from the provided centerline text files.

Missing docstring.

Missing docstring for append_to_curve!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_new_angle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bound_heading. Check Documenter's build log for details.

Missing docstring.

Missing docstring for append_headings. Check Documenter's build log for details.

AutomotiveInteraction.centerlines_txt2tracks — Function

function centerlines_txt2tracks(filename)

Reads a .txt file which contains x coords in col 1 and y coords in col2

Returns a track, i.e. a curve from AutomotiveDrivingModels

Examples

# See make_roadway_interaction()

AutomotiveInteraction.make_roadway_interaction — Function

function make_roadway_interaction()

Overview

Make the DR_CHN_Merging roadway by reading in the centerlines in AutomotiveInteraction.jl/dataset

These centerlines have been modified from the original ones in the centerlines_DR_CHN_Merging_ZS folder

The long lane being merged into by on ramp was split into before and after merge point in 2 separate txt files

Eg: output_centerline_1.txt has been split into centerlines_b1.txt (before merge point) and centerlines_b2.txt (after).

Similarly, output_centerline_5.txt has been split into cetnerlines_g.txt (before) and centerlines_h.txt (after)

Finally, the files in centerlines_DR_CHN_Merging_ZS folder output_centerline_<>.txt had x coords in row 1 and y coords in row 2.

The current centerlines_<>.txt have x in col 1 and y in col 2

Details

Reads in the centerline information from AutomotiveInteraction.jl/dataset

Segment 1: Has 2 lanes: On ramp a is lane 1 and b1 is lane 2

Segment 2: Has 1 lane: b1 is lane 1. Both a and b1 connect into b2

Segment 3: Has 2 lanes: c is lane 1 and d is lane 2

Segment 4: Has 3 lanes: g is lane 1, f1 is lane 2, e1 is lane 3.

Segment 5: Has 2 lanes: f2 is lane 1, e2 is lane 2. e1 connects to e2 and both g and f1 to e2

Note that convention used here is that right most lane in direction of travel is given lane id 1

Examples

roadway_interaction = make_roadway_interaction()

AutomotiveInteraction.make_roadway_ngsim — Function

function make_roadway_ngsim()

Read in the ngsim 101 roadway from the provided roadway file /dataset/ngsim_101.txt

Examples

roadway_ngsim = make_roadway_ngsim()

AutomotiveInteraction.make_roadway_interaction_with_extensions — Function

Make roadway interaction but with added extensions to avoid vehicle getting blocked at road end

Missing docstring.

Missing docstring for make_discont_roadway_straight. Check Documenter's build log for details.

Missing docstring.

Missing docstring for make_discont_roadway_jagged. Check Documenter's build log for details.

AutomotiveInteraction.MergingRoadway — Type

struct MergingRoadway

Wrapper around a roadway for special rendering avoiding lane boundary wierd in merge

Example

road_ext = make_roadway_interaction_with_extensions()
roadway = MergingRoadway(road_ext)
render([roadway])

AutomotiveVisualization.add_renderable! — Function

function AutomotiveVisualization.add_renderable!(rendermodel::RenderModel,mr::MergingRoadway)

Dedicated render function for merging roadway