Visualization

These functions are responsible to create visualizations using AutomotiveVisualization.jl.

AutomotiveInteraction.video_trajdata_replay — Function

function video_trajdata_replay

Makes a video of the trajdata taking frame range as input

Example

video_trajdata_replay(range=1:100,roadway=roadway,trajdata=traj_interaction,
    filename=joinpath(@__DIR__,"../julia_notebooks/media/replay_vid.mp4")

AutomotiveInteraction.scenelist2video — Function

function scenelist2video(scene_list;filename = "media/scenelist_to_video.mp4")

Make video from a list of scenes (generally generated using get_hallucination_scenes

Examples

# See run_vehicles function in driving_simulation.jl

AutomotiveInteraction.video_overlay_scenelists — Function

function videooverlayscenelists(scenelist1,scenelist2,roadway,filename)

Example

scene_list_1 = run_vehicles(id_list=[6,8,19,28,29],roadway=road_ext,traj=traj_ext,
filename="model_driven.mp4")
scene_list_2 = traj_ext[1:length(scene_list_1)]
video_overlay_scenes(scene_list_1,scene_list_2,id_list=[6,8,19,28,29],
roadway=road_ext,filename="model_vs_truth.mp4")

AutomotiveInteraction.scenelist2video_curvepts — Function

function scenelist2video(scene_list;filename = "media/scenelist_to_video.mp4")

Make video from a list of scenes (generally generated using get_hallucination_scenes

Examples

scene = Scene(500)
get!(scene,traj_interaction,1);
scene_list = get_hallucination_scenes(scene,models=models);
scenelist2video(scene_list,filename="media/scenelist_to_video.mp4")