Roadway Building

AutomotiveInteraction.append_to_curve! — Function

function append_to_curve!

AutomotiveInteraction.get_new_angle — Function

function get_new_angle

Does the actual angle calculation based on the x y coordinates

AutomotiveInteraction.bound_heading — Function

function bound_heading

Make the angle range from 0 to pi instead of going beyond

AutomotiveInteraction.append_headings — Function

function append_headings

Used create the angles and append them into the coordinates

Examples

x_coods = [1089.07510, 1093.82626, 1101.19325, 1112.59899, 1123.96733, 1133.24150, 1146.47964]
y_coods = [936.31213, 936.92692,938.52419, 940.93865, 943.27882, 945.21039, 947.88488]
coods = hcat(x_coods,y_coods)
append_headings(coods)

AutomotiveInteraction.centerlines_txt2tracks — Function

function centerlines_txt2tracks(filename)

Reads a .txt file which contains x coords in col 1 and y coords in col2

Returns a track, i.e. a curve from AutomotiveDrivingModels

Examples

# See make_roadway_interaction()

AutomotiveInteraction.make_roadway_interaction — Function

function make_roadway_interaction()

Overview

Make the DR_CHN_Merging roadway by reading in the centerlines in AutomotiveInteraction.jl/dataset

These centerlines have been modified from the original ones in the centerlines_DR_CHN_Merging_ZS folder

The long lane being merged into by on ramp was split into before and after merge point in 2 separate txt files

Eg: output_centerline_1.txt has been split into centerlines_b1.txt (before merge point) and centerlines_b2.txt (after).

Similarly, output_centerline_5.txt has been split into cetnerlines_g.txt (before) and centerlines_h.txt (after)

Finally, the files in centerlines_DR_CHN_Merging_ZS folder output_centerline_<>.txt had x coords in row 1 and y coords in row 2.

The current centerlines_<>.txt have x in col 1 and y in col 2

Details

Reads in the centerline information from AutomotiveInteraction.jl/dataset

Segment 1: Has 2 lanes: On ramp a is lane 1 and b1 is lane 2

Segment 2: Has 1 lane: b1 is lane 1. Both a and b1 connect into b2

Segment 3: Has 2 lanes: c is lane 1 and d is lane 2

Segment 4: Has 3 lanes: g is lane 1, f1 is lane 2, e1 is lane 3.

Segment 5: Has 2 lanes: f2 is lane 1, e2 is lane 2. e1 connects to e2 and both g and f1 to e2

Note that convention used here is that right most lane in direction of travel is given lane id 1

Examples

roadway_interaction = make_roadway_interaction()

Missing docstring.

Missing docstring for make_ngsim_roadway. Check Documenter's build log for details.